Crate typeuri

Named trait for types with globally unique type URIs.

Re-exports

pub use cityhasher;

pub use dashmap;

Traits

Named

A Named type is a type that has a globally unique name.

Derive Macros

Named

Derive the [typeuri::Named] trait for a struct or enum.